moved the random user login to menu
