+++ client/components/organizations.js
@@ -4,7 +4,6 @@ import { connect } from 'react-redux';
 import {fetchOrganizations} from '../store/organizations';
 import { Button, Segment, Divider, Item, Image } from 'semantic-ui-react'
 
-import {createUser} from '../store/user';
 /* -----------------    COMPONENT     ------------------ */
 
 class Organizations extends Component {
@@ -18,9 +17,8 @@ class Organizations extends Component {
       <div>
         {/* {this.user && this.user.isAdmin &&  */}
         <Segment>
-          <Button onClick={this.props.generateRandomUser}>Generate Random User</Button>
           <Button as={Link} to="/organizations/add" floated="right">Add a new Organization</Button>
-          <Divider horizontal>Admin Only</Divider>
+          <Divider horizontal>Staff Only</Divider>
         </Segment>
         <h1>{displayName}</h1>
         {organizations.length > 0 && (
@@ -62,23 +60,6 @@ const mapAllOrganizationsState = ({ user, organizations }) => ({
 
 const mapAllOrganizationsDispatch = dispatch => ({
   loadOrganizations: () => dispatch(fetchOrganizations()),
-  generateRandomUser: () => {
-    fetch('https://randomuser.me/api/')
-    .then(results => results.json())
-    .then(results => {
-      const [randomUser] = results.results;
-      const {name, email, picture, login} = randomUser;
-      const firstName = name.first[0].toUpperCase() + name.first.slice(1);
-      const lastName = name.last[0].toUpperCase() + name.last.slice(1);
-      const newUser = {
-        displayName: firstName + ' ' + lastName,
-        email,
-        photoURL: picture.large,
-        uid: login.md5,
-      }
-      dispatch(createUser(newUser));
-    });
-  }
 });
 
 export const AllOrganizations = connect(mapAllOrganizationsState, mapAllOrganizationsDispatch)(Organizations);
